# Refactor `binary_search_by` to use conditional moves