@pnkfelix 2023-11-29
* The single primary regression here seems to be a measurement blip, based on the 30-day history.
* Even if it weren't, the improvements would outweigh the regression.
* Marked as triaged.

@rustbot label: +perf-regression-triaged